@@ -5,24 +5,31 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node server.js"
+    "start": "node --experimental-specifier-resolution=node index.js"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "body-parser": "^1.19.2",
+    "command-exists": "^1.2.9",
     "cors": "^2.8.5",
+    "dotenv": "^16.0.0",
     "eslint-config-prettier": "^8.5.0",
     "express": "^4.17.3",
+    "multer": "^1.4.4",
     "mysql": "^2.18.1",
-    "prettier": "^2.5.1"
+    "node-fetch": "^2.6.7",
+    "prettier": "^2.5.1",
+    "progress": "^2.0.3"
   },
   "devDependencies": {
-    "eslint": "^8.10.0",
+    "eslint": "^8.13.0",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-plugin-import": "^2.25.4",
     "eslint-plugin-jsx-a11y": "^6.5.1",
     "eslint-plugin-react": "^7.29.3",
     "eslint-plugin-react-hooks": "^4.3.0"
-  }
+  },
+  "type": "module"
 }
