@@ -1,15 +1,7 @@
 module.exports = {
-  env: {
-    browser: true,
-    es2021: true,
-    node: true,
-    jest: true,
-  },
-  extends: ['plugin:react/recommended', 'airbnb', 'prettier'],
-  parser: '@typescript-eslint/parser',
-  parserOptions: {
-    ecmaFeatures: {
-      jsx: true,
+    "env": {
+        "browser": true,
+        "es2021": true
     },
     ecmaVersion: 'latest',
     sourceType: 'module',
@@ -23,8 +15,18 @@ module.exports = {
         endOfLine: 'auto',
       },
     ],
-    semi: 0,
-    'comma-dangle': 0,
-    'react/jsx-filename-extension': [1, { extensions: ['.js', '.jsx'] }],
-  },
-};
+    "parserOptions": {
+        "ecmaFeatures": {
+            "jsx": true
+        },
+        "ecmaVersion": "latest",
+        "sourceType": "module"
+    },
+    "plugins": [
+        "react"
+    ],
+    "rules": {
+        "no-unused-vars": "off",
+        "react/prop-types": "off"
+    }
+}
